requirements.txt
Django==4.2
google_api_python_client
google_apis_oauth_django
